TEST-#2995: fix problems for properties and instancer
